split analyses tile client component
